requirements.txt
fastembed==0.7.4
FlashRank==0.2.10
langchain==1.2.10
langchain_core==1.2.11
langchain_deepseek==1.0.1
langgraph==1.0.8
pandas==3.0.0
pydantic==2.12.5
python-dotenv==1.2.1
qdrant_client==1.16.2
tenacity==9.1.2
tqdm==4.67.1
